backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.2
PyPDF2==3.0.1
python-pptx==0.6.23
pydantic==2.5.0
pdfplumber==0.11.4
PyMuPDF==1.23.5
pytesseract==0.3.10
gunicorn==21.2.0
boto3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
